+++ tests/reveal.ts
@@ -1,0 +1,3 @@
+import { StringUtils } from "../mod.ts";
+
+await StringUtils.reveal("Manually test this and ensure it works.");